backend/requirements.txt
flask==3.1.3
flask-cors==6.0.2
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
matplotlib>=3.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
